export_wiznotes/requirements.txt
requests>=2.25.0
pathlib
markdownify>=0.11.0
websocket-client>=1.0.0
tqdm>=4.60.0